+++ src/components/DashboardFilters/DashboardFilters-container.tsx
@@ -1,14 +1,13 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import Grid from '@material-ui/core/Grid/Grid';
 import Typography from '@material-ui/core/Typography';
 import AutorenewIcon from '@material-ui/icons/Autorenew';
 import clsx from 'clsx';
 
-import FilterDataPicker from './FilterDataPicker-view';
-import useFiltersStyles from './Filters-styles';
+import useFiltersStyles from './DashboardFilters-styles';
+import { Link } from '@material-ui/core';
 
 const Filters = () => {
-  const classes = useFiltersStyles();
 
   const [activeBlockchainButtons, setBlockchainButtons]: [Record<string, boolean>, Function] = useState({
     'BTC': false,
@@ -25,10 +24,10 @@ const Filters = () => {
   const [activeZoomButtons, setZoomButtons]: [Record<string, boolean>, Function] = useState({
     '1 day': true,
     '7 days': false,
-    '1 month': false,
-    '3 months': false,
-    '1 year': false,
-    'All': false
+    // '1 month': false,
+    // '3 months': false,
+    // '1 year': false,
+    // 'All': false
   });
 
   const [activeTopListButtons, setTopListButtons]: [Record<string, boolean>, Function] = useState({
@@ -36,6 +35,10 @@ const Filters = () => {
     '100': false,
     '1000': false
   });
+
+  const classes = useFiltersStyles();
+
+ //const [filters, setFilters]: [string, Function] = useState(' ');
 
   const blockchainFilterHandler = (buttonLabel: string) => {
     if (buttonLabel === 'ETH' || buttonLabel === 'XTZ') {
@@ -65,6 +68,18 @@ const Filters = () => {
     filterType === activeTopListButtons && zoomTopListHandler(buttonLabel);
   };
 
+  const activeFilters = (filterObj: any) => Object.keys(filterObj).filter((item: any) =>  filterObj[item] === true);
+
+  // useEffect((): any => {
+  //   const activeBlockchain = activeFilters(activeBlockchainButtons);
+  //   const activeZoom = activeFilters(activeZoomButtons);
+  //   const activeTopList = activeFilters(activeTopListButtons);
+  //   setFilters({ activeBlockchain, activeZoom, activeTopList});
+
+  // }, [activeBlockchainButtons, activeZoomButtons, activeTopListButtons]);
+
+  
+
   const renderButtons = (buttonLabels: Record<string, boolean>) =>
     Object.keys(buttonLabels).map((buttonLabel: string) => {
       const btnClass = clsx(classes.button, {
@@ -82,6 +97,15 @@ const Filters = () => {
       );
     });
 
+  const handleRefresh = () => {
+    const activeBlockchain = activeFilters(activeBlockchainButtons);
+    const activeZoom = activeFilters(activeZoomButtons);
+    const activeTopList = activeFilters(activeTopListButtons);
+
+    
+    //setFilters({ activeBlockchain, activeZoom, activeTopList});
+  }
+
   return (
     <Grid container justify="flex-start" alignItems="flex-start" className="Container">
       <Grid item xs={3}>
@@ -92,18 +116,15 @@ const Filters = () => {
         <Typography variant="h3">Zoom</Typography>
         {renderButtons(activeZoomButtons)}
       </Grid>
-      <Grid item xs={2}>
-        <Typography variant="h3">Date</Typography>
-        <FilterDataPicker label="From" />
-        <FilterDataPicker label="To" />
-      </Grid>
       <Grid item xs={3}>
         <Typography variant="h3">Top list</Typography>
         {renderButtons(activeTopListButtons)}
       </Grid>
-      <Grid item xs={1}>
+      <Grid item xs={3}>
         <Typography variant="h3">Refresh</Typography>
-        <AutorenewIcon className={classes.refresh} />
+        {/* <Link to={`/filters/${filters}`}> */}
+        <AutorenewIcon className={classes.refresh} onClick={handleRefresh} />
+        {/* </Link> */}
       </Grid>
     </Grid>
   );
