+++ src/components/DashboardFilters/index.tsx
@@ -1,0 +1,3 @@
+import Filters from './DashboardFilters-container';
+
+export default Filters;
