--- src/components/Filters/FilterDataPicker-view.tsx
@@ -1,41 +1,0 @@
-import React from 'react';
-import 'date-fns';
-import Grid from '@material-ui/core/Grid';
-import DateFnsUtils from '@date-io/date-fns';
-import { MuiPickersUtilsProvider,  KeyboardDatePicker} from '@material-ui/pickers';
-
-import useFiltersStyles from './Filters-styles'
-
-const FilterDataPicker = (props: any) => {
-  const [selectedDate, setSelectedDate] = React.useState<Date | null>(
-    new Date(),
-  );
-
-   const classes = useFiltersStyles();
-
-  const handleDateChange = (date: Date | null) => {
-    setSelectedDate(date);
-  };
-
-  return (
-    <MuiPickersUtilsProvider utils={DateFnsUtils}>
-      <Grid container justify="space-around">
-        <KeyboardDatePicker
-          disableToolbar
-          variant="inline"
-          format="dd/MM/yyyy"
-          margin="normal"
-          id="date-picker-inline"
-          label={props.label}
-          value={selectedDate}
-          onChange={handleDateChange}
-          KeyboardButtonProps={{
-            'aria-label': 'change date',
-          }}
-        />
-      </Grid>
-    </MuiPickersUtilsProvider>
-  );
-}
-
-export default FilterDataPicker;
