--- src/components/Filters/index.tsx
@@ -1,3 +1,0 @@
-import Filters from './Filters-container';
-
-export default Filters;
